Moved `PaddedCell.DirtyState` to its own top-level class with new methods.
