+++ CHANGES.md
@@ -12,6 +12,7 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (
 ## [TBD lint release]
 * **BREAKING** Removed `isClean`, `applyTo`, and `applyToAndReturnResultIfDirty` from `Formatter` because users should instead use `PaddedCell.check()`.
 * **BREAKING** Removed `FormatterStep.Strict` because it was unnecessary and unused implementation detail.
+* **BREAKING** Moved `PaddedCell.DirtyState` to its own top-level class with new methods.
 
 ## [Unreleased]
 ### Changes
