+++ lib/src/main/java/com/diffplug/spotless/PaddedCell.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2022 DiffPlug
+ * Copyright 2016-2023 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,18 +18,13 @@ package com.diffplug.spotless;
 import static com.diffplug.spotless.LibPreconditions.requireElementsNonNull;
 
 import java.io.File;
-import java.io.IOException;
-import java.io.OutputStream;
 import java.nio.file.Files;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
 import java.util.function.Function;
-
-import javax.annotation.Nullable;
 
 /**
  * Models the result of applying a {@link Formatter} on a given {@link File}
@@ -176,108 +171,4 @@ public final class PaddedCell {
 		}
 		// @formatter:on
 	}
-
-	/**
-	 * Calculates whether the given file is dirty according to a PaddedCell invocation of the given formatter.
-	 * DirtyState includes the clean state of the file, as well as a warning if we were not able to apply the formatter
-	 * due to diverging idempotence.
-	 */
-	public static DirtyState calculateDirtyState(Formatter formatter, File file) throws IOException {
-		Objects.requireNonNull(formatter, "formatter");
-		Objects.requireNonNull(file, "file");
-
-		byte[] rawBytes = Files.readAllBytes(file.toPath());
-		return calculateDirtyState(formatter, file, rawBytes);
-	}
-
-	public static DirtyState calculateDirtyState(Formatter formatter, File file, byte[] rawBytes) throws IOException {
-		String raw = new String(rawBytes, formatter.getEncoding());
-		// check that all characters were encodable
-		String encodingError = EncodingErrorMsg.msg(raw, rawBytes, formatter.getEncoding());
-		if (encodingError != null) {
-			throw new IllegalArgumentException(encodingError);
-		}
-		String rawUnix = LineEnding.toUnix(raw);
-
-		// enforce the format
-		String formattedUnix = formatter.compute(rawUnix, file);
-		// convert the line endings if necessary
-		String formatted = formatter.computeLineEndings(formattedUnix, file);
-
-		// if F(input) == input, then the formatter is well-behaving and the input is clean
-		byte[] formattedBytes = formatted.getBytes(formatter.getEncoding());
-		if (Arrays.equals(rawBytes, formattedBytes)) {
-			return isClean;
-		}
-
-		// F(input) != input, so we'll do a padded check
-		String doubleFormattedUnix = formatter.compute(formattedUnix, file);
-		if (doubleFormattedUnix.equals(formattedUnix)) {
-			// most dirty files are idempotent-dirty, so this is a quick-short circuit for that common case
-			return new DirtyState(formattedBytes);
-		}
-
-		PaddedCell cell = PaddedCell.check(formatter, file, rawUnix);
-		if (!cell.isResolvable()) {
-			return didNotConverge;
-		}
-
-		// get the canonical bytes
-		String canonicalUnix = cell.canonical();
-		String canonical = formatter.computeLineEndings(canonicalUnix, file);
-		byte[] canonicalBytes = canonical.getBytes(formatter.getEncoding());
-		if (!Arrays.equals(rawBytes, canonicalBytes)) {
-			// and write them to disk if needed
-			return new DirtyState(canonicalBytes);
-		} else {
-			return isClean;
-		}
-	}
-
-	/**
-	 * The clean/dirty state of a single file.  Intended use:
-	 * - {@link #isClean()} means that the file is clean, and there's nothing else to say
-	 * - {@link #didNotConverge()} means that we were unable to determine a clean state
-	 * - once you've tested the above conditions and you know that it's a dirty file with a converged state,
-	 *   then you can call {@link #writeCanonicalTo(OutputStream)} to get the canonical form of the given file.
-	 */
-	public static class DirtyState {
-		@Nullable
-		private final byte[] canonicalBytes;
-
-		private DirtyState(@Nullable byte[] canonicalBytes) {
-			this.canonicalBytes = canonicalBytes;
-		}
-
-		public boolean isClean() {
-			return this == isClean;
-		}
-
-		public boolean didNotConverge() {
-			return this == didNotConverge;
-		}
-
-		private byte[] canonicalBytes() {
-			if (canonicalBytes == null) {
-				throw new IllegalStateException("First make sure that {@code !isClean()} and {@code !didNotConverge()}");
-			}
-			return canonicalBytes;
-		}
-
-		public void writeCanonicalTo(File file) throws IOException {
-			Files.write(file.toPath(), canonicalBytes());
-		}
-
-		public void writeCanonicalTo(OutputStream out) throws IOException {
-			out.write(canonicalBytes());
-		}
-	}
-
-	/** Returns the DirtyState which corresponds to {@code isClean()}. */
-	public static DirtyState isClean() {
-		return isClean;
-	}
-
-	private static final DirtyState didNotConverge = new DirtyState(null);
-	private static final DirtyState isClean = new DirtyState(null);
 }
