+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/SpotlessTaskImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2022 DiffPlug
+ * Copyright 2016-2023 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -36,8 +36,8 @@ import org.gradle.work.FileChange;
 import org.gradle.work.InputChanges;
 
 import com.diffplug.common.base.StringPrinter;
+import com.diffplug.spotless.DirtyState;
 import com.diffplug.spotless.Formatter;
-import com.diffplug.spotless.PaddedCell;
 import com.diffplug.spotless.extra.GitRatchet;
 
 @CacheableTask
@@ -95,11 +95,11 @@ public abstract class SpotlessTaskImpl extends SpotlessTask {
 	private void processInputFile(@Nullable GitRatchet ratchet, Formatter formatter, File input) throws IOException {
 		File output = getOutputFile(input);
 		getLogger().debug("Applying format to " + input + " and writing to " + output);
-		PaddedCell.DirtyState dirtyState;
+		DirtyState dirtyState;
 		if (ratchet != null && ratchet.isClean(getProjectDir().get().getAsFile(), getRootTreeSha(), input)) {
-			dirtyState = PaddedCell.isClean();
+			dirtyState = DirtyState.clean();
 		} else {
-			dirtyState = PaddedCell.calculateDirtyState(formatter, input);
+			dirtyState = DirtyState.of(formatter, input).calculateDirtyState();
 		}
 		if (dirtyState.isClean()) {
 			// Remove previous output if it exists
